Image transformation feature
- revised code
- moving it outside the classification feature
