+++ app/templates/base.html
@@ -34,6 +34,9 @@
 				<li class="nav-item">
 					<a class="nav-link" href="/histogram">Histogram</a>
 				</li>
+                <li class="nav-item">
+					<a class="nav-link" href="/transformations">Image transformation</a>
+				</li>
             </ul>
         </div>
     </div>
