+++ app/templates/classification_select.html
@@ -12,18 +12,6 @@
             {{ form.model }}
         </p>
         <h4>
-            Transformation:
-        </h4>
-        <p>
-            Color {{ form.color }}
-            <br>
-            Brightness {{ form.brightness }}
-            <br>
-            Contrast {{ form.contrast }}
-            <br>
-            Sharpness {{ form.sharpness }}
-        </p>
-        <h4>
             Image:
         </h4>
         <p>
