+++ app/templates/transformation_select.html
@@ -1,0 +1,28 @@
+{% extends "base.html" %}
+
+{% block content %}
+
+    <h1>Select Image and Transformations</h1>
+    <form action="" method="post" novalidate>
+        {{ form.hidden_tag() }}
+        <h4>
+            Image:
+        </h4>
+        <p>
+            {{ form.image }}
+        </p>
+        <h4>
+            Transformation:
+        </h4>
+        <p>
+            Color {{ form.color }}
+            <br>
+            Brightness {{ form.brightness }}
+            <br>
+            Contrast {{ form.contrast }}
+            <br>
+            Sharpness {{ form.sharpness }}
+        </p>
+        <button type="submit" class="btn btn-dark mb-2">Submit</button>
+    </form>
+{% endblock %}
